Animate a depth dependent timeseries from a Dlf file

Animate a depth dependent timeseries from a Dlf file

Usage

animate_depth_timeseries(dlf, var_name)

Arguments

dlf

An S4 object of class Dlf

var_name

Name of variable in columns

Value

A plotly figure

Examples

data_dir <- system.file("extdata", package="daisyrVis")
path <- file.path(data_dir, "daily/DailyP/DailyP-Daily-WaterFlux.dlf")
dlf <- read_dlf(path)
#> Warning: number of items read is not a multiple of the number of columns
animate_depth_timeseries(dlf, "q")
#> Warning: Ignoring 3 observations