Transform daisy time columns to a single timestamp

Transform daisy time columns to a single timestamp

Usage

daisy_time_to_timestamp(
  dlf,
  time_col_name = "time",
  year_col = NULL,
  month_col = NULL,
  day_col = NULL,
  hour_col = NULL,
  drop_daisy_time_cols = FALSE
)

Arguments

dlf

An S4 object of class Dlf or a list of Dlf objects

time_col_name

The name of the new timestamp column

year_col

Name of year column, set to NULL if no year column

month_col

Name of month column, set to NULL if no month column

day_col

Name of day column, set to NULL if no day column

hour_col

Name of hour column, set to NULL if no hour column

drop_daisy_time_cols

If TRUE drop the year, month, day, hour columns

Value

An S4 object of class Dlf

If all of year_col, month_col, day_col and hour_col are NULL, then an attempt is made to guess the relevant names.

The returned object is identical to the dlf, except that the Columns 'year', 'month', 'mday' and 'hour' are dropped and replaced with a column containing the corresponding POSIXct timestamp

Examples

data_dir <- system.file("extdata", package="daisyrVis")
path <- file.path(data_dir, 'daily/DailyP/DailyP-Daily-WaterFlux.dlf')
dlf <- read_dlf(path)
#> Warning while reading /home/runner/work/_temp/Library/daisyrVis/extdata/daily/DailyP/DailyP-Daily-WaterFlux.dlf
#> Data section starts at row 17
#> Discarded single-line footer: <<1990	4	28	0	0.892003	0.965204	0.992118	0.984267	0.995394	0.867143	0.665696	0.405031	0.360616	0.317983	0.256743	0.121373	0.00275476	-0.0963126	-0.175119	-0.255377	-0.33114	-0.403575	-0.444362	-0.485611	-0.51228	-0.53814	-0.557847	-0.571283	-0.578562	-0.580094>>
dlf@data[1,]
#>   z       time        q
#> 1 0 1990-04-02 0.686972
dlf <- daisy_time_to_timestamp(dlf)
dlf@data[1,]
#>   z       time        q
#> 1 0 2024-11-28 0.686972