Transform wide format time series to a long format time series It is assumed that dlf@data only contains time_name columns and depth columns with the format <var_name>_ e.g. "q_100"

Transform wide format time series to a long format time series It is assumed that dlf@data only contains time_name columns and depth columns with the format <var_name>_ e.g. "q_100"

Usage

depth_wide_to_long(
  dlf,
  var_name,
  time_name = "time",
  depth_name = "z",
  depth_unit = "unknown"
)

Arguments

dlf

An S4 object of class Dlf or a list of Dlf objects

var_name

Name of variable in columns

time_name

Name of time column

depth_name

Name to use for new depth column

depth_unit

Unit of depth

Value

An S4 object of class Dlf

The data of rhe returned Dlf object contains one row for each time/depth combination.

Examples

data_dir <- system.file("extdata", package="daisyrVis")
path <- file.path(data_dir, "daily/DailyP/DailyP-Daily-WaterFlux.dlf")
dlf <- read_dlf(path, convert_depth=FALSE)
#> Warning while reading /home/runner/work/_temp/Library/daisyrVis/extdata/daily/DailyP/DailyP-Daily-WaterFlux.dlf
#> Data section starts at row 17
#> Discarded single-line footer: <<1990	4	28	0	0.892003	0.965204	0.992118	0.984267	0.995394	0.867143	0.665696	0.405031	0.360616	0.317983	0.256743	0.121373	0.00275476	-0.0963126	-0.175119	-0.255377	-0.33114	-0.403575	-0.444362	-0.485611	-0.51228	-0.53814	-0.557847	-0.571283	-0.578562	-0.580094>>
dlf@data[1,]
#>   year month mday hour q..AT..0 q..AT.._1 q..AT.._2 q..AT.._3 q..AT.._5.5
#> 1 1990     4    2    0 0.686972  0.664115   0.63981  0.613606    0.544964
#>   q..AT.._10.5 q..AT.._17 q..AT.._25 q..AT.._27 q..AT.._29 q..AT.._32
#> 1     0.395473   0.175839  -0.137307  -0.191751  -0.246702  -0.330268
#>   q..AT.._41 q..AT.._51 q..AT.._61 q..AT.._70 q..AT.._80 q..AT.._90 q..AT.._100
#> 1  -0.559167  -0.754909  -0.884317  -0.942759    -0.9329  -0.843949   -0.702253
#>   q..AT.._106 q..AT.._113 q..AT.._120 q..AT.._130 q..AT.._140 q..AT.._150
#> 1   -0.617684   -0.542156   -0.506693   -0.485376   -0.474339   -0.468794
#>   q..AT.._160 q..AT.._170 q..AT.._180 q..AT.._190 q..AT.._200       time
#> 1   -0.465219   -0.461336   -0.455971    -0.44889   -0.440634 1990-04-02
dlf <- depth_wide_to_long(dlf, "q")
head(dlf@data)
#>   z       time        q
#> 1 0 1990-04-02 0.686972
#> 2 0 1990-04-03 0.329720
#> 3 0 1990-04-04 0.569315
#> 4 0 1990-04-05 0.552182
#> 5 0 1990-04-06 0.804888
#> 6 0 1990-04-07 1.570560