Transform daisy time columns to a single timestamp

Transform daisy time columns to a single timestamp

Usage

daisy_time_to_timestamp(
  dlf,
  time_col_name = "time",
  year_col = "year",
  month_col = "month",
  day_col = "mday",
  hour_col = "hour",
  drop_daisy_time_cols = FALSE
)

Arguments

dlf

An S4 object of class Dlf or a list of Dlf objects

time_col_name

The name of the new timestamp column

year_col

Name of year column, set to NULL if no year column

month_col

Name of month column, set to NULL if no month column

day_col

Name of day column, set to NULL if no day column

hour_col

Name of hour column, set to NULL if no hour column

drop_daisy_time_cols

If TRUE drop the year, month, day, hour columns

Value

An S4 object of class Dlf

The returned object is identical to the dlf, except that the Columns 'year', 'month', 'mday' and 'hour' are dropped and replaced with a column containing the corresponding POSIXct timestamp

Examples

data_dir <- system.file("extdata", package="daisyrVis")
path <- file.path(data_dir, 'daily/DailyP/DailyP-Daily-WaterFlux.dlf')
dlf <- read_dlf(path)
#> Warning: number of items read is not a multiple of the number of columns
dlf@data[1,]
#>   year month mday hour q..AT..0 q..AT.._1 q..AT.._2 q..AT.._3 q..AT.._5.5
#> 1 1990     4    2    0 0.686972  0.664115   0.63981  0.613606    0.544964
#>   q..AT.._10.5 q..AT.._17 q..AT.._25 q..AT.._27 q..AT.._29 q..AT.._32
#> 1     0.395473   0.175839  -0.137307  -0.191751  -0.246702  -0.330268
#>   q..AT.._41 q..AT.._51 q..AT.._61 q..AT.._70 q..AT.._80 q..AT.._90 q..AT.._100
#> 1  -0.559167  -0.754909  -0.884317  -0.942759    -0.9329  -0.843949   -0.702253
#>   q..AT.._106 q..AT.._113 q..AT.._120 q..AT.._130 q..AT.._140 q..AT.._150
#> 1   -0.617684   -0.542156   -0.506693   -0.485376   -0.474339   -0.468794
#>   q..AT.._160 q..AT.._170 q..AT.._180 q..AT.._190 q..AT.._200
#> 1   -0.465219   -0.461336   -0.455971    -0.44889   -0.440634
dlf <- daisy_time_to_timestamp(dlf)
dlf@data[1,]
#>   year month mday hour q..AT..0 q..AT.._1 q..AT.._2 q..AT.._3 q..AT.._5.5
#> 1 1990     4    2    0 0.686972  0.664115   0.63981  0.613606    0.544964
#>   q..AT.._10.5 q..AT.._17 q..AT.._25 q..AT.._27 q..AT.._29 q..AT.._32
#> 1     0.395473   0.175839  -0.137307  -0.191751  -0.246702  -0.330268
#>   q..AT.._41 q..AT.._51 q..AT.._61 q..AT.._70 q..AT.._80 q..AT.._90 q..AT.._100
#> 1  -0.559167  -0.754909  -0.884317  -0.942759    -0.9329  -0.843949   -0.702253
#>   q..AT.._106 q..AT.._113 q..AT.._120 q..AT.._130 q..AT.._140 q..AT.._150
#> 1   -0.617684   -0.542156   -0.506693   -0.485376   -0.474339   -0.468794
#>   q..AT.._160 q..AT.._170 q..AT.._180 q..AT.._190 q..AT.._200       time
#> 1   -0.465219   -0.461336   -0.455971    -0.44889   -0.440634 1990-04-02